+++ plugins/woocommerce-admin/client/analytics/components/report-chart/index.js
@@ -183,7 +183,7 @@ ReportChart.propTypes = {
 	/**
 	 * Primary data to display in the chart.
 	 */
-	primaryData: PropTypes.object.isRequired,
+	primaryData: PropTypes.object,
 	/**
 	 * The query string represented in object form.
 	 */
@@ -198,10 +198,33 @@ ReportChart.propTypes = {
 	selectedChart: PropTypes.object.isRequired,
 };
 
+ReportChart.defaultProps = {
+	primaryData: {
+		data: {
+			intervals: [],
+		},
+		isError: false,
+		isRequesting: false,
+	},
+	secondaryData: {
+		data: {
+			intervals: [],
+		},
+		isError: false,
+		isRequesting: false,
+	},
+};
+
 export default compose(
 	withSelect( ( select, props ) => {
 		const { query, endpoint, filters } = props;
 		const chartMode = props.mode || getChartMode( filters, query ) || 'time-comparison';
+
+		if ( query.search && ! ( query[ endpoint ] && query[ endpoint ].length ) ) {
+			return {
+				mode: chartMode,
+			};
+		}
 
 		if ( 'item-comparison' === chartMode ) {
 			const primaryData = getReportChartData( endpoint, 'primary', query, select );