+++ plugins/woocommerce-admin/client/wc-api/wc-api-spec.js
@@ -5,7 +5,6 @@
  */
 import items from './items';
 import notes from './notes';
-import products from './products';
 import reportItems from './reports/items';
 import reportStats from './reports/stats';
 import reviews from './reviews';
@@ -21,7 +20,6 @@ function createWcApiSpec() {
 		selectors: {
 			...items.selectors,
 			...notes.selectors,
-			...products.selectors,
 			...reportItems.selectors,
 			...reportStats.selectors,
 			...reviews.selectors,
@@ -33,7 +31,6 @@ function createWcApiSpec() {
 				return [
 					...items.operations.read( resourceNames ),
 					...notes.operations.read( resourceNames ),
-					...products.operations.read( resourceNames ),
 					...reportItems.operations.read( resourceNames ),
 					...reportStats.operations.read( resourceNames ),
 					...reviews.operations.read( resourceNames ),