+++ plugins/woocommerce-admin/includes/data-stores/class-wc-admin-reports-taxes-data-store.php
@@ -257,7 +257,9 @@ class WC_Admin_Reports_Taxes_Data_Store extends WC_Admin_Reports_Data_Store impl
 	public static function sync_order_taxes( $order_id ) {
 		global $wpdb;
 		$order = wc_get_order( $order_id );
-		if ( ! $order ) {
+
+		// Skip `shop_order_refunds` when factoring stats on order tax.
+		if ( ! $order || 'shop_order' !== $order->get_type() ) {
 			return -1;
 		}