+++ plugins/woocommerce-admin/packages/components/src/chart/constants.js
@@ -1,0 +1,13 @@
+/** @format */
+
+// This is the max number of items that can be selected/shown on a chart at one time.
+// If this number changes, the color scale also needs to be adjusted.
+export const selectionLimit = 5;
+export const colorScales = [
+	[],
+	[ 0.5 ],
+	[ 0.333, 0.667 ],
+	[ 0.25, 0.50, 0.75 ],
+	[ 0.20, 0.40, 0.60, 0.80 ],
+	[ 0.16, 0.32, 0.48, 0.64, 0.80 ],
+];