+++ plugins/woocommerce-admin/packages/components/src/chart/d3chart/utils/color.js
@@ -1,25 +1,18 @@
 /** @format */
 
 /**
+ * Internal dependencies
+ */
+import { colorScales, selectionLimit } from '../../constants';
+
+/**
  * External dependencies
  */
 import { findIndex } from 'lodash';
 
 export const getColor = ( key, orderedKeys, colorScheme ) => {
-	const smallColorScales = [
-		[],
-		[ 0.5 ],
-		[ 0.333, 0.667 ],
-		[ 0.2, 0.5, 0.8 ],
-		[ 0.12, 0.375, 0.625, 0.88 ],
-	];
-	let keyValue = 0;
-	const len = orderedKeys.length;
+	const len = orderedKeys.length > selectionLimit ? selectionLimit : orderedKeys.length;
 	const idx = findIndex( orderedKeys, d => d.key === key );
-	if ( len < 5 ) {
-		keyValue = smallColorScales[ len ][ idx ];
-	} else {
-		keyValue = idx / ( orderedKeys.length - 1 );
-	}
+	const keyValue = idx <= ( selectionLimit - 1 ) ? colorScales[ len ][ idx ] : 0;
 	return colorScheme( keyValue );
 };