+++ plugins/woocommerce-admin/packages/components/src/chart/index.js
@@ -24,6 +24,7 @@ import { updateQueryString } from '@woocommerce/navigation';
 import ChartPlaceholder from './placeholder';
 import { H, Section } from '../section';
 import { D3Chart, D3Legend } from './d3chart';
+import { selectionLimit } from './constants';
 
 function getD3CurrencyFormat( symbol, position ) {
 	switch ( position ) {
@@ -74,7 +75,7 @@ function getOrderedKeys( props, previousOrderedKeys = [] ) {
 			return updatedKeys.filter( key => key.total > 0 ).map( ( key, index ) => {
 				return {
 					...key,
-					visible: index < 5 || key.visible,
+					visible: index < selectionLimit || key.visible,
 				};
 			} );
 		}
@@ -105,15 +106,15 @@ class Chart extends Component {
 	}
 
 	componentDidUpdate( prevProps ) {
-		const { data, query, mode } = this.props;
+		const { data, query, isRequesting, mode } = this.props;
 		if ( ! isEqual( [ ...data ].sort(), [ ...prevProps.data ].sort() ) ) {
 			/**
 			 * Only update the orderedKeys when data is present so that
 			 * selection may persist while requesting new data.
 			 */
-			const orderedKeys = data.length
-				? getOrderedKeys( this.props, this.state.orderedKeys )
-				: this.state.orderedKeys;
+			const orderedKeys = isRequesting && ! data.length
+				? this.state.orderedKeys
+				: getOrderedKeys( this.props, this.state.orderedKeys );
 			/* eslint-disable react/no-did-update-set-state */
 			this.setState( {
 				orderedKeys,