+++ plugins/woocommerce-admin/packages/navigation/src/history.js
@@ -6,4 +6,13 @@ import { createHashHistory } from 'history';
 
 // See https://github.com/ReactTraining/react-router/blob/master/FAQ.md#how-do-i-access-the-history-object-outside-of-components
 
-export default createHashHistory();
+let _history;
+
+function getHistory() {
+	if ( ! _history ) {
+		_history = createHashHistory();
+	}
+	return _history;
+}
+
+export { getHistory };