+++ plugins/woocommerce-admin/packages/navigation/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.0.0
+
+- Replace `history` export with `getHistory` (allows for lazy-create of history)
+
 # 1.1.0
 
 - Rename `getTimeRelatedQuery` to `getPersistedQuery`